requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
async-timeout==4.0.3
atpublic==5.1
attrs==25.3.0
bigframes==2.19.0
boto3==1.40.22
botocore==1.40.22
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
cloudpickle==3.1.1
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
db-dtypes==1.4.3
debugpy==1.8.16
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
filetype==1.2.0
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.9.0
gcsfs==2025.9.0
geopandas==1.1.1
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-cloud-bigquery==3.37.0
google-cloud-bigquery-connection==1.18.3
google-cloud-bigquery-storage==2.33.1
google-cloud-core==2.4.3
google-cloud-functions==1.20.4
google-cloud-resource-manager==1.14.2
google-cloud-storage==3.3.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.13.0
idna==3.10
implicit==0.7.2
ipykernel==6.30.1
ipython==8.37.0
ipywidgets==8.1.7
jedi==0.19.2
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
langchain==0.3.27
langchain-aws==0.2.31
langchain-core==0.3.75
langchain-google-genai==2.1.10
langchain-text-splitters==0.3.11
langsmith==0.4.21
lightfm==1.17
llvmlite==0.44.0
markdown-it-py==4.0.0
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.6.4
nest-asyncio==1.6.0
numba==0.61.2
numpy==1.26.4
nvidia-nccl-cu12==2.27.7
oauthlib==3.3.1
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pandas-gbq==0.29.2
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydata-google-auth==1.9.1
Pygments==2.19.2
pyogrio==0.11.1
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9.1
s3transfer==0.13.1
scikit-learn==1.7.1
scipy==1.14.1
seaborn==0.13.2
shap==0.48.0
shapely==2.1.1
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlglot==27.13.2
stack-data==0.6.3
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
toolz==1.0.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
widgetsnbextension==4.0.14
xgboost==3.0.4
yarl==1.20.1
zstandard==0.24.0
